requirements.txt
python-dotenv==1.0.1
slack_bolt==1.20.1
slack_sdk==3.33.1
